+++ decidim-assemblies-extended/app/decorators/forms/decidim/assemblies/admin/assembly_form_decorator.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+Decidim::Assemblies::Admin::AssemblyForm.class_eval do
+  attribute :show_home, Virtus::Attribute::Boolean
+end
